# pos-web/Swagger/pos-api-request.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "POS.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Costs": {
      "get": {
        "tags": [
          "Costs"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Costs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Cost.CreateCostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Cost.CreateCostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Cost.CreateCostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
    "/api/Costs/{id}": {
      "get": {
        "tags": [
          "Costs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Costs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Cost.UpdateCostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Cost.UpdateCostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Cost.UpdateCostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Cost.CostDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Costs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Object, System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Object, System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Object, System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "POS.Application.DTOs.Cost.CreateCostDto": {
        "type": "object",
        "properties": {
          "productVariantId": {
            "type": "string",
            "format": "uuid"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "costValue": {
            "type": "number",
            "format": "double"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "POS.Application.DTOs.Cost.UpdateCostDto": {
        "type": "object",
        "properties": {
          "costValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter 'Bearer' [space] and then your valid token in the text input below.\r\n\r\nExample: \"Bearer eyJhbGciOiJIUzI1Ni...\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}